+++ test/pixel/Jamfile
@@ -17,4 +17,7 @@ compile concepts.cpp ;
 compile is_pixel.cpp ;
 compile is_planar.cpp ;
 compile num_channels.cpp ;
+compile pixel_reference_is_mutable.cpp ;
 compile pixels_are_compatible.cpp ;
+
+run test_fixture.cpp ;
